@@ -10,7 +10,7 @@ homepage = "https://apps.gnome.org/PikaBackup/"
 repository = "https://gitlab.gnome.org/World/pika-backup"
 edition = "2021"
 default-run = "pika-backup"
-rust-version = "1.76"
+rust-version = "1.80"
 
 [profile.release]
 lto = true
@@ -39,7 +39,6 @@ zxcvbn = "2.0"
 arc-swap = "1.0"
 enclose = "1.1"
 itertools = "0.13"
-once_cell = "1.3"
 quick-error = "2.0"
 zeroize = "1.1"
 
