@@ -1,19 +1,19 @@
 use crate::config;
 pub use crate::globals::*;
 use arc_swap::ArcSwap;
-use once_cell::sync::Lazy;
 use std::collections::HashMap;
 use std::rc::Rc;
+use std::sync::LazyLock;
 
-pub static BACKUP_CONFIG: Lazy<ArcSwap<config::Backups>> = Lazy::new(Default::default);
-pub static BACKUP_HISTORY: Lazy<ArcSwap<config::Histories>> = Lazy::new(Default::default);
+pub static BACKUP_CONFIG: LazyLock<ArcSwap<config::Backups>> = LazyLock::new(Default::default);
+pub static BACKUP_HISTORY: LazyLock<ArcSwap<config::Histories>> = LazyLock::new(Default::default);
 
-pub static SCHEDULE_STATUS: Lazy<ArcSwap<config::Writeable<config::ScheduleStatus>>> =
-    Lazy::new(Default::default);
+pub static SCHEDULE_STATUS: LazyLock<ArcSwap<config::Writeable<config::ScheduleStatus>>> =
+    LazyLock::new(Default::default);
 
 /// Last reminded about not meeting criteria
-pub static LAST_REMINDED: Lazy<ArcSwap<HashMap<config::ConfigId, std::time::Instant>>> =
-    Lazy::new(Default::default);
+pub static LAST_REMINDED: LazyLock<ArcSwap<HashMap<config::ConfigId, std::time::Instant>>> =
+    LazyLock::new(Default::default);
 
 thread_local!(
     static GIO_APPLICATION: Rc<gio::Application> = Rc::new({
