@@ -4,8 +4,8 @@ use crate::ui::prelude::*;
 use glib::prelude::*;
 use glib::subclass::prelude::*;
 use glib::{ParamSpec, ParamSpecString};
-use once_cell::sync::Lazy;
 use std::cell::RefCell;
+use std::sync::LazyLock;
 
 #[derive(Debug, Clone)]
 pub enum PrunePreset {
@@ -94,8 +94,8 @@ mod imp {
 
     impl ObjectImpl for PrunePresetObject {
         fn properties() -> &'static [ParamSpec] {
-            static PROPERTIES: Lazy<Vec<ParamSpec>> =
-                Lazy::new(|| vec![ParamSpecString::builder("display").build()]);
+            static PROPERTIES: LazyLock<Vec<ParamSpec>> =
+                LazyLock::new(|| vec![ParamSpecString::builder("display").build()]);
             PROPERTIES.as_ref()
         }
 
